// packages/validation/src/api-gen/zod.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

import { z } from 'zod';

export const zCurrentResponse = z.object({
    email: z.string(),
    name: z.string(),
    pid: z.string()
});

export const zForgotParams = z.object({
    email: z.email()
});

export const zLoginParams = z.object({
    email: z.email(),
    password: z.string()
});

export const zLoginResponse = z.object({
    is_verified: z.boolean(),
    name: z.string(),
    pid: z.string(),
    token: z.string()
});

export const zMagicLinkParams = z.object({
    email: z.email()
});

export const zRegisterParams = z.object({
    email: z.email(),
    name: z.string().min(2),
    password: z.string()
});

export const zResendVerificationParams = z.object({
    email: z.email()
});

export const zResetParams = z.object({
    password: z.string(),
    token: z.string()
});

export const zCurrentData = z.object({
    body: z.optional(z.never()),
    path: z.optional(z.never()),
    query: z.optional(z.never())
});

/**
 * Current user data
 */
export const zCurrentResponse2 = zCurrentResponse;

export const zForgotData = z.object({
    body: zForgotParams,
    path: z.optional(z.never()),
    query: z.optional(z.never())
});

export const zLoginData = z.object({
    body: zLoginParams,
    path: z.optional(z.never()),
    query: z.optional(z.never())
});

/**
 * Login successful
 */
export const zLoginResponse2 = zLoginResponse;

export const zMagicLinkData = z.object({
    body: zMagicLinkParams,
    path: z.optional(z.never()),
    query: z.optional(z.never())
});

export const zMagicLinkVerifyData = z.object({
    body: z.optional(z.never()),
    path: z.object({
        token: z.string()
    }),
    query: z.optional(z.never())
});

/**
 * Magic link login successful
 */
export const zMagicLinkVerifyResponse = zLoginResponse;

export const zRegisterData = z.object({
    body: zRegisterParams,
    path: z.optional(z.never()),
    query: z.optional(z.never())
});

export const zResendVerificationEmailData = z.object({
    body: zResendVerificationParams,
    path: z.optional(z.never()),
    query: z.optional(z.never())
});

export const zResetData = z.object({
    body: zResetParams,
    path: z.optional(z.never()),
    query: z.optional(z.never())
});

export const zVerifyData = z.object({
    body: z.optional(z.never()),
    path: z.object({
        token: z.string()
    }),
    query: z.optional(z.never())
});